Rename component for clarity
