+++ src/components/pages/callouts/CalloutThanksBox.vue
@@ -1,16 +1,20 @@
 <template>
-  <AppBoxout :title="callout.thanksTitle" :icon="faThumbsUp" variant="success">
+  <AppMessageBox
+    :title="callout.thanksTitle"
+    :icon="faThumbsUp"
+    variant="success"
+  >
     <div
       class="content-message font-normal text-body-80"
       v-html="callout.thanksText"
     />
-  </AppBoxout>
+  </AppMessageBox>
 </template>
 
 <script lang="ts" setup>
 import { faThumbsUp } from '@fortawesome/free-solid-svg-icons';
 import { GetCalloutDataWith } from '../../../utils/api/api.interface';
-import AppBoxout from '../../AppBoxout.vue';
+import AppMessageBox from '../../AppMessageBox.vue';
 
 defineProps<{
   callout: GetCalloutDataWith<'form'>;
